TST/REF: collect tests by method, some misplaced (#37354)
